Support multiple network devices in benchmark tool
Summary:
Systems such as the DGX-1 have more than one network adapter. This
change lets the benchmark tool use them. This option must be used
together with the --threads option. Threads pick up devices to use in a
round-robin fashion. If you run with 2 devices but with a 1 thread,
you'll still only run on 1 device; the first one. It is good practice to
use a number of threads equal to an integer multiple of the number of
devices.
Closes https://github.com/facebookincubator/gloo/pull/68

Differential Revision: D5586439

Pulled By: pietern

fbshipit-source-id: 394f2320752c56cd34eb813fcdfb334494c2e4e8