Use localhost interface for all addr in localhost subnet
Summary: sockaddrToInterfaceName() fails when the resolved address for gethostname() is not 127.0.0.1 (but sitll in the localhost subnet). This change ensures that all addresses in the localhost subnet map to the localhost interface. I do not generically map addresses to interface using the interface netmask because we may have multiple interfaces for the same subnet, however, there is only one interface for localhost.

Reviewed By: pietern

Differential Revision: D5531536

fbshipit-source-id: 9a4f8822086467b48e9432e345b2c4c5b55bd9e0