Make CUDA algorithms use stream associated with scratch pointer
Summary:
These algorithms still defaulted to the stream of the first pointer.
When GPUDirect is used, we'll use a scratch buffer that is closest to
the networking adapter to minimize intra-machine communication.
However, since streams are pinned to devices, we have to use the
stream associated to the device that hosts the scratch buffer.
Closes https://github.com/facebookincubator/gloo/pull/69

Reviewed By: wesolwsk

Differential Revision: D5736512

Pulled By: pietern

fbshipit-source-id: 4ef4a01835670d16ef253ade986f73cfb32177d0