Stop relying on bundled version of NCCL
Summary:
NCCL1 is deprecated in favor of NCCL2 and if anyone wants to still use
it they can specify to do so via NCCL_ROOT_DIR.

This change removes the NCCL submodule and makes use of NCCL a compile
time option and constant. If USE_NCCL=ON, then all device side
reductions are performed using NCCL. Otherwise, they are performed with
a tree of CUDA reductions. Calls to cudaDeviceReduce or
cudaDeviceBroadcast are handled based on the USE_NCCL constant.
Closes https://github.com/facebookincubator/gloo/pull/80

Reviewed By: wesolwsk

Differential Revision: D5746687

Pulled By: pietern

fbshipit-source-id: fa527ae72ddf5b4fe3497239789376d6a6f99b33