Step 1 of phased migration of timeout to context
Summary:
The timeout "state" is currently associated with the device instance
and carried over to pairs upon creating them. It requires the API user
to call setTimeout on the device before creating new pairs. This is
cumbersome when you need to clone a context because there is no
natural way to carry over its timeout. This commit is phase 1 in a 2
phase migration to store the timeout with the context instead of the
device.

Reviewed By: andrewwdye

Differential Revision: D5769247

fbshipit-source-id: ed6601ac532b60c52cd85f7b9412ac9580a56e00