Add AllreduceBuilder algorithm instance builder
Summary:
This builder class is an initial stab at moving the decision which
algorithm implementation to use to Gloo. Different machine or network
configurations, or buffer sizes, warrant different implementations.
Hardcoding these decisions over and over in downstream code is not
productive. Instead, we can (eventually) have Gloo make these
decisions for us.

Another benefit of deferring this decision is that we can more easily
choose the optimal implementation in case there is just a single
process with multiple buffers. Downstream code currently has to make a
distinction between single machine and multi machine configurations
and removing this distinction means these code paths can be
simplified.

This commit only contains a builder for Allreduce. The benchmark code
only uses the builder in case it runs an allreduce benchmark and the
current approach otherwise. When builders for the other collectives
are added we should up the benchmark main code as well.
Closes https://github.com/facebookincubator/gloo/pull/90

Reviewed By: romain-intel

Differential Revision: D5925628

Pulled By: pietern

fbshipit-source-id: 1380db31cde4e3ace53c458628b734a3e04379d1